… site/source/docs/api_reference/emscripten.h.rst
… site/source/docs/api_reference/fiber.h.rst
… site/source/docs/api_reference/index.rst
… src/settings.js
… src/struct_info.json
… system/include/emscripten/emscripten.h
… system/include/emscripten/fiber.h
… tests/test_core.py
… tests/test_fibers.cpp