As discussed in #8979, this supersedes coroutines. Implemented for the Upstream backend + Asyncify only.

The API is slightly different from that of the [libkoishi backend](https://github.com/taisei-project/koishi/blob/b16c82fc6e86a1623d8ea553bfb934aee05033dc/src/asyncify/library_fiber.js) which it's based on. The main difference is that the fiber structure does not contain a stack. The user must allocate memory for a stack manually before creating a fiber.

Tests and documentation are still missing, I'll be adding it shortly.

cc @kripken @corwin-of-amber @sbc100