@Akaricchi 2019-11-22
@kripken looks like I've found a much better solution to the stack overflow in 1dd209b97203ff6eac043f4297b11ffdc9e85ce1. Let me know what you think.

Also, do you have any suggestions as to where and how I should get started with the tests? The `tests` directory looks pretty intimidating.