@kripken 2019-11-27
Nice, I like that solution! :smiley: 

Can we use that to get rid of the existing `afterUnwind` cases? (As the two mechanisms overlap, but your new one is more general.)

For tests, yeah, there are quite a lot... but for our purposes here, I'd say we can just look at the ones for fastcomp coroutines, and create new tests parallel to those. Looks like there are two existing tests, with names starting with `test_coroutine*` in `tests/test_core.py`, both using `tests/test_coroutines.cpp`. So maybe adding `tests/test_fibers.cpp`, and a new test `test_fibers()` right after those coroutine ones in `test_core.py`.