@Akaricchi 2019-11-28
>Can we use that to get rid of the existing afterUnwind cases? (As the two mechanisms overlap, but your new one is more general.)

There's a possibility, but I'm not entirely sure. The new mechanism is slightly different in that the trampoline callback is invoked only after the outermost `handleSleep` is unwound, whereas `afterUnwind` runs after every nested one as well. I'm not sure whether I fully understand the implications of that outside of fiber code. It'd need some testing for sure. Perhaps we should leave them separate for the purposes of this PR, and try to reconcile them later after it's in?