@juj 2019-11-29
Can all of this go into a separate `emscripten/fiber.h` header, and use a separate `src/library_fiber.js` file for the implementation as well? That would give a better organization to these kind of distinct feature blocks.