@Akaricchi 2019-11-29
Yes, that shouldn't be a problem. I simply followed the coroutines example when I decided where to put it.