@kripken 2019-12-03
> Perhaps we should leave them separate for the purposes of this PR, and try to reconcile them later after it's in?

Sounds reasonable, sure.