@Akaricchi 2019-12-03
> I think I'm missing something important here. When you call swap, it saves Asyncify.currData and then after unwinding it switches Asyncify.currData to the one we want to rewind into. But how did we get that second data in the first place?

When we enter a freshly created fiber, there is nothing to rewind into, and `Asyncify.currData` becomes 0. The fiber's continuation function in this case is the one set up by `emscripten_fiber_create` or `emscripten_fiber_recycle`. When we leave that fiber later, `emscripten_fiber_swap` will update its `currData` pointer and continuation (to the `wakeUp` callback), so that it can be properly rewound into later if we need to re-enter the fiber.