@Akaricchi 2019-12-09
Following @sbc100's advice, I rewrote this to represent fibers as JS objects instead of on-heap structs. I'm quite satisfied with the results; the code is now more readable and the libkoishi benchmark now actually runs a bit faster. `emscripten_fiber` is now really an integer handle and not a real pointer, however I left it typedef'd to `struct emscripten_fiber *` for some basic type safety.