@Akaricchi 2019-12-16
Actually, don't merge it just yet... I've discovered some serious optimization opportunities. Will be updating it shortly.