@kripken 2019-12-17
I think it's ok to leave optimization for followup work. Especially if it adds any complexity, it's better to have as little as possible in the initial landing. But up to you.