@Akaricchi 2019-12-24
Reverted the instrumentation optimization, let's see if CI likes it now...

Benchmark before reverting:
```
10000000 switches in 3.124312s  ~3200704.610795 switches/s      ~312 ns/switch
10000000 switches in 3.197342s  ~3127597.407095 switches/s      ~319 ns/switch
10000000 switches in 3.182314s  ~3142367.479135 switches/s      ~318 ns/switch
10000000 switches in 3.158110s  ~3166450.915520 switches/s      ~315 ns/switch
10000000 switches in 3.100296s  ~3225498.399868 switches/s      ~310 ns/switch
10000000 switches in 3.116179s  ~3209058.272968 switches/s      ~311 ns/switch
10000000 switches in 3.233076s  ~3093029.582124 switches/s      ~323 ns/switch
10000000 switches in 3.164263s  ~3160293.505066 switches/s      ~316 ns/switch
10000000 switches in 3.236837s  ~3089435.429198 switches/s      ~323 ns/switch
10000000 switches in 3.302496s  ~3028012.696554 switches/s      ~330 ns/switch
```

Benchmark after reverting:
```
10000000 switches in 3.813196s  ~2622471.939443 switches/s      ~381 ns/switch
10000000 switches in 3.723655s  ~2685533.786891 switches/s      ~372 ns/switch
10000000 switches in 3.821741s  ~2616608.729760 switches/s      ~382 ns/switch
10000000 switches in 3.774961s  ~2649033.710027 switches/s      ~377 ns/switch
10000000 switches in 3.763688s  ~2656968.395474 switches/s      ~376 ns/switch
10000000 switches in 3.759762s  ~2659742.985439 switches/s      ~375 ns/switch
10000000 switches in 3.724611s  ~2684844.319936 switches/s      ~372 ns/switch
10000000 switches in 3.737705s  ~2675438.708895 switches/s      ~373 ns/switch
10000000 switches in 3.714831s  ~2691912.858992 switches/s      ~371 ns/switch
10000000 switches in 3.791190s  ~2637694.346615 switches/s      ~379 ns/switch
```

Note that the performance has apparently increased overall since the last time I measured it. It's probably because of a large system update I've installed recently.